Removed Optional Reshade Files
As stated in a forum post ( https://reshade.me/forum/general-discussion/771-distribution-regulations ) the distribution of the reshade binaries is allowed. Distribution of the shaders are their assets is not under jurisdiction of the reshade author but instead the author of the shaders.

"As said before I'm personally ok with the distribution of the ReShade binaries, but since the shaders are not mine, I have no right to allow that for them too, that's up to their authors."

Due to this the shaders are being removed alongside the compatibility files from dgVoodoo.

A tutorial .txt or .md file of how to get reshade working with dgVoodoo might possibly take the place of the drag and drop solution.